createOrUpdateGameForm view for creating and updating game

+++ src/main/java/org/springframework/samples/SevenIslands/general/GeneralService.java
@@ -31,12 +31,12 @@ public class GeneralService {
 
     @Transactional
     public int totalDurationAll(){
-        return (int) generalRepo.totalDurationAllGames();
+        return generalRepo.totalDurationAllGames();
     }
 
     @Transactional
     public int totalGamesCount(){
-        return (int) generalRepo.totalGames();
+        return generalRepo.totalGames();
     }
 
     public void insertIdUser(Map<String, Object> model){
